requirements.txt
aiohttp>=3.9.0
httpx>=0.27.0
loguru>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
telethon>=1.35.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
